# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Job scheduling and task management service for Minisource",
        "title": "Scheduler Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5003",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/jobs": {
            "get": {
                "description": "List jobs with optional filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "List jobs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_minisource_scheduler_internal_models.Job"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new scheduled job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Create a job",
                "parameters": [
                    {
                        "description": "Job creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_minisource_scheduler_internal_models.CreateJobRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_minisource_scheduler_internal_models.Job"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/jobs/stats": {
            "get": {
                "description": "Get statistics about jobs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get job statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_minisource_scheduler_internal_models.JobStats"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/jobs/{id}": {
            "get": {
                "description": "Get a job by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get a job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_minisource_scheduler_internal_models.Job"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Update a job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Job update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_minisource_scheduler_internal_models.UpdateJobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_minisource_scheduler_internal_models.Job"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Soft-delete a job",
                "tags": [
                    "jobs"
                ],
                "summary": "Delete a job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/jobs/{id}/pause": {
            "post": {
                "description": "Pause a job from executing",
                "tags": [
                    "jobs"
                ],
                "summary": "Pause a job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_minisource_scheduler_internal_models.Job"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/jobs/{id}/resume": {
            "post": {
                "description": "Resume a paused job",
                "tags": [
                    "jobs"
                ],
                "summary": "Resume a job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_minisource_scheduler_internal_models.Job"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/jobs/{id}/trigger": {
            "post": {
                "description": "Manually trigger a job execution",
                "tags": [
                    "jobs"
                ],
                "summary": "Trigger a job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_minisource_scheduler_internal_models.JobExecution"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_minisource_scheduler_internal_models.CreateJobRequest": {
            "type": "object",
            "required": [
                "endpoint",
                "name",
                "schedule",
                "type"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "headers": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "max_retries": {
                    "type": "integer"
                },
                "metadata": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "method": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "priority": {
                    "type": "integer"
                },
                "retry_delay": {
                    "type": "integer"
                },
                "schedule": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "timeout": {
                    "type": "integer"
                },
                "timezone": {
                    "type": "string"
                },
                "type": {
                    "enum": [
                        "cron",
                        "one_time",
                        "interval"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_minisource_scheduler_internal_models.JobType"
                        }
                    ]
                }
            }
        },
...
        "github_com_minisource_scheduler_internal_models.Job": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endpoint": {
                    "description": "HTTP endpoint to call",
                    "type": "string"
                },
                "fail_count": {
                    "type": "integer"
                },
                "headers": {
                    "description": "HTTP headers",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "string"
                },
                "last_run_at": {
                    "type": "string"
                },
                "max_retries": {
                    "description": "Max retry attempts",
                    "type": "integer"
                },
                "metadata": {
                    "description": "Additional metadata",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "method": {
                    "description": "HTTP method",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "next_run_at": {
                    "type": "string"
                },
                "payload": {
                    "description": "Request body",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "priority": {
                    "description": "1-10, higher is more important",
                    "type": "integer"
                },
                "retry_delay": {
                    "description": "Delay between retries in seconds",
                    "type": "integer"
                },
                "run_count": {
                    "type": "integer"
                },
                "schedule": {
                    "description": "Cron expression or interval",
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_minisource_scheduler_internal_models.JobStatus"
                },
                "tags": {
                    "description": "Job tags for filtering",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "tenant_id": {
                    "type": "string"
                },
                "timeout": {
                    "description": "Timeout in seconds",
                    "type": "integer"
                },
                "timezone": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_minisource_scheduler_internal_models.JobType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_minisource_scheduler_internal_models.JobStats": {
            "type": "object",
            "properties": {
                "active_jobs": {
                    "type": "integer"
                },
                "avg_duration_ms": {
                    "type": "number"
                },
                "failures_today": {
                    "type": "integer"
                },
                "jobs_by_status": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "jobs_by_type": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "paused_jobs": {
                    "type": "integer"
                },
                "runs_today": {
                    "type": "integer"
                },
                "success_rate": {
                    "type": "number"
                },
                "total_jobs": {
                    "type": "integer"
                },
                "total_runs": {
                    "type": "integer"
                }
            }
        },
...
        "internal_handler.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/internal_handler.ErrorInfo"
                },
                "meta": {
                    "$ref": "#/definitions/internal_handler.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}